refactor-openhands: Rename remotes.js -> remote.js
- lib/remotes.js -> lib/remote.js
- Update 7 imports in lib/branch.js, lib/connectors/*, lib/sync.js
