+++ lib/branch.js
@@ -15,7 +15,7 @@
  */
 
 // Try to use provider, fallback to CLI
-const { getProvider, detectProvider } = require('./remotes');
+const { getProvider, detectProvider } = require('./remote');
 
 let provider = null;
 let useProvider = false;
