+++ lib/connectors/bitbucket.js
@@ -1,4 +1,4 @@
-class BitbucketProvider extends require("../remotes").GitProvider {
+class BitbucketProvider extends require("../remote").GitProvider {
     constructor(config = {}) {
         super({ ...config, type: 'bitbucket' });
         this.token = config.token || process.env.BITBUCKET_TOKEN;
