+++ lib/connectors/selfhosted.js
@@ -1,4 +1,4 @@
-class SelfHostedProvider extends require("../remotes").GitProvider {
+class SelfHostedProvider extends require("../remote").GitProvider {
     constructor(config = {}) {
         super({ ...config, type: 'selfhosted' });
         this.url = config.url || config.remoteUrl || 'origin';
