+++ lib/sync.js
@@ -189,7 +189,7 @@ function getAllCircuits() {
  * @returns {GitProvider[]} Array of configured providers
  */
 function getConfiguredProviders() {
-    const { getAllProviders } = require('./remotes');
+    const { getAllProviders } = require('./remote');
     const providers = getAllProviders();
     
     const configured = [];
@@ -372,7 +372,7 @@ function isRAID() {
  * @returns {object} Rebase result
  */
 async function rebase(providerName) {
-    const { getAllProviders } = require('./remotes');
+    const { getAllProviders } = require('./remote');
     const providers = getAllProviders();
     const provider = providers[providerName];
